update report stats

+++ elasticsearch.go
@@ -112,7 +112,7 @@ func sendToES(config *elasticConfig, log *logrus.Entry, stats *counters, batch [
 			log.Warn("Failed to post batch")
 		} else {
 			log.Debug("Completed post")
-			stats.esSent += len(batch)
+			stats.esSent += int64(len(batch))
 			stats.batchesSent++
 		}
 	}
